requirements.txt
aiogram==3.8.0
Telethon==1.36.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
python-dotenv==1.0.1
apscheduler==3.10.4
openai==1.43.0
tenacity==8.5.0
tiktoken==0.7.0
uvloop==0.19.0
orjson==3.10.7